refactoring code structure
