+++ src/exceptions/base_exception.py
@@ -1,0 +1,4 @@
+
+
+class BasePhonebookException(Exception):
+    """Базовое исключение для всех ошибок"""
