+++ src/utils.py
@@ -1,6 +1,6 @@
-import settings
-import text
-import view
+from src import settings
+from src import text
+from src import view
 
 
 def ask_yes_no(question: str, default: str = "yes"):
